\name{DisenoVikua-package}
\alias{DisenoVikua-package}
\docType{package}
\title{
\packageTitle{DisenoVikua}
}
\description{
\packageDescription{DisenoVikua}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{DisenoVikua}
\packageIndices{DisenoVikua}
~~ An overview of how to use the package, including the most important ~~
~~ functionalias{DisenoVikua}
}
\author{
\packageAuthor{DisenoVikua}
Maintainer: \packageMaintainer{DisenoVikua}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}