\name{PaquetePrueba-package}
\alias{PaquetePrueba-package}
{{{\docType{package}
 \title{
 \packageTitle{PaquetePrueba}
 }
 \description{
 \packageDescription{PaquetePrueba}
 }
 \details{

 The DESCRIPTION file:
 \packageDESCRIPTION{PaquetePrueba}
 \packageIndices{PaquetePrueba}
 ~~ An overview of how to use the package, including the most important ~~
 ~~ functionalias{PaquetePrueba}
\s ~~
}
\author{
\packageAuthor{PaquetePrueba}

Maintainer: \packageMaintainer{PaquetePrueba}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
